move `IsReapable` behaviour into `common-lib` (as `ReapableEligibility` trait) and ignore reapable images during migration
