+++ media-api/app/lib/MediaApiConfig.scala
@@ -50,11 +50,6 @@ class MediaApiConfig(resources: GridConfigResources) extends CommonConfigWithEla
   val persistenceIdentifier = string("persistence.identifier")
   val queriableIdentifiers = Seq(persistenceIdentifier)
 
-  val persistedRootCollections: List[String] = stringOpt("persistence.collections") match {
-    case Some(collections) => collections.split(',').toList
-    case None => List(s"${staffPhotographerOrganisation} Archive")
-  }
-
   def convertToInt(s: String): Option[Int] = Try { s.toInt }.toOption
 
   val syndicationStartDate: Option[DateTime] = Try {
