+++ thrall/app/ThrallComponents.scala
@@ -53,7 +53,17 @@ class ThrallComponents(context: Context) extends GridComponents(context, new Thr
 
   val uiSource: Source[KinesisRecord, Future[Done]] = KinesisSource(highPriorityKinesisConfig)
   val automationSource: Source[KinesisRecord, Future[Done]] = KinesisSource(lowPriorityKinesisConfig)
-  val migrationSourceWithSender: MigrationSourceWithSender = MigrationSourceWithSender(materializer, auth.innerServiceCall, es, gridClient, config.projectionParallelism)
+  val migrationSourceWithSender: MigrationSourceWithSender = MigrationSourceWithSender(
+    materializer,
+    auth.innerServiceCall,
+    es,
+    gridClient,
+    config.projectionParallelism,
+    isReapableQuery = new ReapableEligibility {
+      val persistedRootCollections: List[String] = config.persistedRootCollections
+      val maybePersistenceIdentifier: Option[String] = config.maybePersistenceIdentifier
+    }.query
+  )
 
   val thrallEventConsumer = new ThrallEventConsumer(
     es,
