+++ thrall/app/lib/MigrationSourceWithSender.scala
@@ -7,7 +7,8 @@ import com.gu.mediaservice.GridClient
 import com.gu.mediaservice.lib.elasticsearch.{InProgress, Paused}
 import com.gu.mediaservice.lib.logging.GridLogging
 import com.gu.mediaservice.model.{MigrateImageMessage, MigrationMessage}
-import lib.elasticsearch.{ElasticSearch, ScrolledSearchResults}
+import com.sksamuel.elastic4s.requests.searches.queries.Query
+import lib.elasticsearch.{ElasticSearch, ReapableEligibility, ScrolledSearchResults}
 import play.api.libs.ws.WSRequest
 
 import java.time.Instant
@@ -29,6 +30,7 @@ object MigrationSourceWithSender extends GridLogging {
     es: ElasticSearch,
     gridClient: GridClient,
     projectionParallelism: Int,
+    isReapableQuery: Query
   )(implicit ec: ExecutionContext): MigrationSourceWithSender = {
 
     // scroll through elasticsearch, finding image ids and versions to migrate
@@ -65,7 +67,7 @@ object MigrationSourceWithSender extends GridLogging {
             val nextIdsToMigrate = ((es.migrationStatus, maybeScrollId) match {
               case (Paused(_), _) => Future.successful(List.empty)
               case (InProgress(migrationIndexName), None) =>
-                es.startScrollingImageIdsToMigrate(migrationIndexName).map(handleScrollResponse)
+                es.startScrollingImageIdsToMigrate(migrationIndexName, isReapableQuery).map(handleScrollResponse)
               case (InProgress(_), Some(scrollId)) =>
                 es.continueScrollingImageIdsToMigrate(scrollId).map(handleScrollResponse)
               case _ => Future.successful(List.empty)
