+++ thrall/test/lib/ThrallStreamProcessorTest.scala
@@ -10,8 +10,9 @@ import com.gu.mediaservice.GridClient
 import com.gu.mediaservice.lib.aws.UpdateMessage
 import com.gu.mediaservice.lib.json.JsonByteArrayUtil
 import com.gu.mediaservice.model.{MigrateImageMessage, StaffPhotographer, ThrallMessage}
+import com.sksamuel.elastic4s.ElasticApi.matchNoneQuery
 import helpers.Fixtures
-import lib.elasticsearch.{ElasticSearch, ScrolledSearchResults}
+import lib.elasticsearch.{ElasticSearch, ReapableEligibility, ScrolledSearchResults}
 import lib.kinesis.ThrallEventConsumer
 import org.mockito.ArgumentMatchers.any
 import org.mockito.Mockito.when
@@ -109,7 +110,7 @@ class ThrallStreamProcessorTest extends AnyFunSpec with BeforeAndAfterAll with M
     lazy val mockEs = mock[ElasticSearch]
     when(mockEs.continueScrollingImageIdsToMigrate(any())(any(), any()))
       .thenReturn(Future.successful(ScrolledSearchResults(List.empty, None)))
-    when(mockEs.startScrollingImageIdsToMigrate(any())(any(), any()))
+    when(mockEs.startScrollingImageIdsToMigrate(any(), any())(any(), any()))
     .thenReturn(Future.successful(ScrolledSearchResults(List.empty, None)))
 
     val uiPrioritySource: Source[KinesisRecord, Future[Done.type]] =
@@ -121,7 +122,8 @@ class ThrallStreamProcessorTest extends AnyFunSpec with BeforeAndAfterAll with M
       (req: WSRequest) => req,
       mockEs,
       mockGrid,
-      projectionParallelism = 1
+      projectionParallelism = 1,
+      isReapableQuery = matchNoneQuery()
     )
 
     lazy val mockConsumer: ThrallEventConsumer = mock[ThrallEventConsumer]
